Separated tunefeedforward into its own cpp file. Added a comment explaining what all is going on
